% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci}
\alias{get_ci}
\title{This calculates an upper and lower bound from bootstrap matrix}
\usage{
get_ci(grid_val,
       coef_mat,
       degree)
}
\arguments{
\item{grid_val}{is the vector of grid values on \code{treat} axis}

\item{coef_mat}{contains the bootstrapped parameter estimates.}

\item{degree}{is 1 for linear and 2 for quadratic outcome model}
}
\value{
\code{get_ci} returns upper and lower 95 percent confidence lines.
}
\description{
This function takes a matrix containing the bootstrapped coefficients
from a parametric ADRF estimator and returns upper and lower 95 percent
confidence lines.
}

