% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hi_sim_data.R
\docType{data}
\name{hi_sim_data}
\alias{hi_sim_data}
\title{Simulated data from Hirano and Imbens (2004)}
\format{A data frame with 1000 rows and 6 variables:}
\source{
use the \code{hi_sample} function
}
\usage{
data(hi_sim_data)
}
\description{
Simulated data used in the paper "The propensity score with continuous
treatments."
}
\details{
A dataset containing hi_sim_data.
}
\examples{
## Example from Hirano and Imbens (2004).
data(hi_sim_data)
head(hi_sim_data)
}
\references{
Hirano, Keisuke, and Guido W. Imbens.
"The propensity score with continuous treatments."
\emph{Applied Bayesian modeling and causal inference from
incomplete-data perspectives} (2004): 73-84.


Moodie, Erica EM, and David A. Stephens.
 "Estimation of dose-response functions for longitudinal data
  using the generalised propensity score."
  \emph{Statistical methods in medical research} \bold{21.2} (2012): 149-166.
}
\keyword{datasets}

