% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nmes_data.R
\docType{data}
\name{nmes_data}
\alias{nmes_data}
\title{Data set containing data from the National Medical Expenditure Survey (NMES)}
\format{A dataset containing 9708 observations and 12 variables.}
\usage{
data(nmes_data)
}
\description{
Data set from the NMES.
with 9708 observations and 12 variables.
}
\examples{
data(nmes_data)
head(nmes_data)
}
\references{
Imai, K., & van Dyk, D.A. (2004).
Causal Inference With General Treatment Regimes: Generalizing the Propensity Score.
\emph{Journal of the American Statistical Association}, \bold{99(467)}.


National Center for Health Services Research and Health Care Technology Assessment.
NATIONAL MEDICAL EXPENDITURE SURVEY, 1987:
INSTITUTIONAL POPULATION COMPONENT. Rockville, MD: Westat, Inc. [producer], 1987.
 Ann Arbor, MI: Inter-university Consortium for Political and
 Social Research [distributor], 1990. doi:10.3886/ICPSR09280.v1

Bryer, Jason M.
"TriMatch: An R Package for Propensity Score Matching of Non-binary Treatments."
The R User Conference, useR! 2013 July 10-12 2013
University of Castilla-La Mancha, Albacete, Spain. Vol. 10. No. 30. 2013.
}
\keyword{datasets}

